refactored image upload
